@@ -3,37 +3,56 @@
   "version": "1.0.0",
   "main": "./node_modules/expo/AppEntry.js",
   "scripts": {
-    "start": "expo start",
+    "start": "expo start --tunnel",
     "android": "expo start --android",
     "ios": "expo start --ios",
     "web": "expo start --web"
   },
   "dependencies": {
+    "@expo/config": "^8.1.1",
+    "@expo/metro-config": "^0.10.0",
+    "@react-native-async-storage/async-storage": "^1.23.1",
+    "@react-native-community/datetimepicker": "7.7.0",
+    "@react-native-community/geolocation": "^3.2.1",
+    "@react-native-picker/picker": "^2.7.5",
     "@react-navigation/bottom-tabs": "^6.5.19",
     "@react-navigation/native": "^6.1.16",
     "@react-navigation/native-stack": "^6.9.25",
     "@react-navigation/stack": "^6.3.28",
     "@rneui/base": "^4.0.0-rc.7",
     "@rneui/themed": "^4.0.0-rc.8",
-    "@shopify/react-native-skia": "^1.0.2",
-    "@types/react": "~18.2.14",
+    "@shopify/react-native-skia": "1.2.3",
+    "@types/react": "~18.2.79",
+    "axios": "^1.6.8",
     "d3": "^7.9.0",
-    "expo": "^49.0.23",
-    "expo-barcode-scanner": "~12.5.3",
-    "expo-camera": "~13.4.4",
-    "expo-image-picker": "~14.3.2",
-    "expo-status-bar": "~1.11.1",
+    "date-fns": "^3.6.0",
+    "expo": "~51.0.4",
+    "expo-barcode-scanner": "~13.0.1",
+    "expo-camera": "~15.0.9",
+    "expo-image-picker": "~15.0.4",
+    "expo-location": "~17.0.1",
+    "expo-status-bar": "~1.12.1",
+    "formik": "^2.4.6",
     "nativewind": "^2.0.11",
     "react": "18.2.0",
-    "react-native": "0.73.4",
+    "react-native": "0.74.1",
     "react-native-calendars": "^1.1304.1",
     "react-native-camera": "^4.2.1",
     "react-native-chart-kit": "^6.12.0",
+    "react-native-circular-progress-indicator": "^4.4.2",
+    "react-native-gifted-charts": "^1.4.10",
+    "react-native-linear-gradient": "^2.8.3",
+    "react-native-maps": "1.14.0",
+    "react-native-maps-directions": "^1.9.0",
+    "react-native-progress": "^5.0.1",
+    "react-native-radio-buttons-group": "^3.1.0",
+    "react-native-redash": "^18.1.3",
     "react-native-responsive-screen": "^1.4.2",
-    "react-native-safe-area-context": "4.8.2",
-    "react-native-screens": "~3.29.0",
-    "react-native-svg": "^15.1.0",
-    "typescript": "^5.1.3"
+    "react-native-safe-area-context": "4.10.1",
+    "react-native-screens": "3.31.1",
+    "react-native-svg": "^15.3.0",
+    "typescript": "~5.3.3",
+    "yup": "^1.4.0"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
