@@ -1,20 +1,12 @@
 import { SafeAreaProvider } from "react-native-safe-area-context";
-import { NavigationContainer } from "@react-navigation/native";
-import TabNavigation from "./App/Navigations/TabNavigation";
-import { View, Text } from "react-native";
-import AuthStack from "./App/Navigations/AuthStack";
-import WelcomeScreen from "./App/Screens/Authentication/WelcomeScreen";
+
+import { AuthProvider } from "./App/context/AuthContext";
+import AppNav from "./App/Navigations/AppNav";
 function App() {
   return (
-    //SafeAreaProvider is an expo component => prevents the elements to be on the phone nav bar
-    <NavigationContainer>
-      <AuthStack />
-
-      {/*
-            <TabNavigation />
-
-      */}
-    </NavigationContainer>
+    <AuthProvider>
+      <AppNav />
+    </AuthProvider>
   );
 }
 export default App;
