@@ -1,63 +1,153 @@
-import { Image, Text, ScrollView, TouchableOpacity, View } from "react-native";
 import React, { useState } from "react";
+import {
+  Image,
+  Text,
+  ScrollView,
+  TouchableOpacity,
+  View,
+  ActivityIndicator,
+} from "react-native";
+import { useNavigation, useFocusEffect } from "@react-navigation/native";
+import AsyncStorage from "@react-native-async-storage/async-storage";
+import api from "../../Api/api";
 import backgroundImage from "../../../assets/background.png";
 import {
   widthPercentageToDP as wp,
   heightPercentageToDP as hp,
 } from "react-native-responsive-screen";
 import { theme } from "../../Theme";
-import { useNavigation } from "@react-navigation/native";
 
 export default function Slider() {
   const navigation = useNavigation();
+  const [userToken, setUserToken] = useState("");
+  const [email, setEmail] = useState("");
+  const [registration, setRegistration] = useState([]);
+  const [firstRegistration, setFirstRegistration] = useState(null);
+  const [loading, setLoading] = useState(true);
+
+  const fetchRegistration = async () => {
+    const storedToken = await AsyncStorage.getItem("Token");
+    const storedEmail = await AsyncStorage.getItem("Email");
+    setUserToken(storedToken);
+    setEmail(storedEmail);
+    try {
+      const runsResponse = await api.get(
+        `ParticipantRuns/GetRunsByParticipant?email=${storedEmail}`,
+        {
+          headers: { Authorization: `Bearer ${storedToken}` },
+        }
+      );
+      setRegistration(runsResponse.data);
+      if (runsResponse.data.length > 0) {
+        setFirstRegistration(runsResponse.data[0]);
+      }
+      console.log(runsResponse.data);
+    } catch (error) {
+      console.error("Error fetching runs:", error);
+    } finally {
+      setLoading(false);
+    }
+  };
+  useFocusEffect(
+    React.useCallback(() => {
+      fetchRegistration();
+    }, [])
+  );
+  if (loading) {
+    return (
+      <View style={{ flex: 1, justifyContent: "center", alignItems: "center" }}>
+        <ActivityIndicator size="large" color={theme.text} />
+      </View>
+    );
+  }
 
   return (
     <View className="px-5">
       <View className="space-y-5">
-        <View className=" flex-row justify-between items-center">
+        <View className="flex-row justify-between items-center">
           <Text
             style={{ fontSize: wp(5) }}
             className="font-bold text-neutral-700"
           >
-            Future Events
+            Registered Events
           </Text>
           <TouchableOpacity>
             <Text
               style={{ fontSize: wp(4), color: theme.text }}
-              onPress={() => navigation.navigate("Registration")}
+              onPress={() =>
+                navigation.navigate("Registration", {
+                  eventData: registration,
+                })
+              }
             >
               See All
             </Text>
           </TouchableOpacity>
         </View>
       </View>
+      <View>
+        <ScrollView horizontal showsHorizontalScrollIndicator={false}>
+          {firstRegistration ? (
+            <TouchableOpacity
+              onPress={() =>
+                navigation.navigate("Participant run", {
+                  startDate: firstRegistration.startDate.split("T")[0],
+                  startTime: firstRegistration.startTime.substring(0, 5),
+                  eventType: firstRegistration.eventType,
+                  id: firstRegistration.eventAttributeFK,
+                })
+              }
+              style={{ flexDirection: "row", marginTop: 10, marginBottom: 10 }}
+            >
+              <Image
+                source={{ uri: firstRegistration.imageSrc }}
+                style={{
+                  width: wp(40),
+                  height: wp(24),
+                  borderRadius: 10,
+                  borderWidth: 1,
+                }}
+              />
+              <View style={{ marginLeft: 10 }}>
+                <Text style={{ fontSize: wp(4), fontWeight: "bold" }}>
+                  {firstRegistration.eventName}
+                </Text>
+                <Text style={{ fontSize: wp(3), color: theme.text }}>
+                  {firstRegistration.startDate.split("T")[0]} at{" "}
+                  {firstRegistration.startTime.substring(0, 5)}
+                </Text>
+              </View>
+            </TouchableOpacity>
+          ) : (
+            <View
+              style={{
+                justifyContent: "center",
+                alignItems: "center",
+                height: wp(24),
+              }}
+            >
+              <TouchableOpacity
+                style={{
+                  justifyContent: "center",
+                  alignItems: "center",
+                }}
+              >
+                <Text
+                  style={{
+                    fontSize: wp(4.5),
+                  }}
+                >
+                  There is no registration yet
+                </Text>
+              </TouchableOpacity>
+            </View>
+          )}
+        </ScrollView>
+      </View>
 
-      <TouchableOpacity
-        className="flex-row space-y-2 mt-2 mb-2"
-        onPress={() => navigation.navigate("Participant run")}
-      >
-        <Image
-          source={backgroundImage}
-          style={{ width: wp(40), height: wp(24) }}
-          className="rounded-xl"
-        />
-        <View className="flex mx-4">
-          <Text
-            className="text-neutral-700 font-bold mb-2"
-            style={{ fontSize: wp(4), width: wp(40) }}
-            numberOfLines={2}
-          >
-            L- Mobile Marathon Event
-          </Text>
-          <Text className="text-neutral-500 " style={{ fontSize: wp(3) }}>
-            7:00 PM
-          </Text>
-        </View>
-      </TouchableOpacity>
-
-      {/*Try to add multiple images*/}
-      <View className="space-y-5">
-        <View className=" flex-row justify-between items-center">
+      {/*Events
+      <View className="space-y-2">
+        <View className="flex-row justify-between items-center">
           <Text
             style={{ fontSize: wp(5) }}
             className="font-bold text-neutral-700"
@@ -76,21 +166,21 @@ export default function Slider() {
         className="space-x-4"
         showsHorizontalScrollIndicator={false}
       >
-        <TouchableOpacity className="flex items-center space-y-2 ">
+        <TouchableOpacity className="flex items-center space-y-2">
           <Image
             source={backgroundImage}
             style={{ width: wp(35), height: wp(34) }}
             className="rounded-xl"
           />
         </TouchableOpacity>
-        <TouchableOpacity className="flex items-center space-y-2 ">
+        <TouchableOpacity className="flex items-center space-y-2">
           <Image
             source={backgroundImage}
             style={{ width: wp(35), height: wp(34) }}
             className="rounded-xl"
           />
         </TouchableOpacity>
-        <TouchableOpacity className="flex items-center space-y-2 ">
+        <TouchableOpacity className="flex items-center space-y-2">
           <Image
             source={backgroundImage}
             style={{ width: wp(35), height: wp(34) }}
@@ -98,6 +188,7 @@ export default function Slider() {
           />
         </TouchableOpacity>
       </ScrollView>
+      */}
     </View>
   );
 }
