@@ -1,194 +1,159 @@
-import { View, Text, Image, ScrollView, TouchableOpacity } from "react-native";
-import React, { useEffect } from "react";
+import React, { useEffect, useRef } from "react";
+import { View, Text, TouchableOpacity, ScrollView } from "react-native";
 import { useRoute } from "@react-navigation/native";
-
-import maps from "../../../assets/maps.png";
+import { StatusBar } from "expo-status-bar";
+import { SafeAreaView } from "react-native-safe-area-context";
+import { Feather } from "@expo/vector-icons";
 import {
   widthPercentageToDP as wp,
   heightPercentageToDP as hp,
 } from "react-native-responsive-screen";
-import { StatusBar } from "expo-status-bar";
-import { SafeAreaView } from "react-native-safe-area-context";
-import { Feather } from "@expo/vector-icons";
 import { theme } from "../../Theme";
+import MapView, { Marker } from "react-native-maps";
+import MapViewDirections from "react-native-maps-directions";
 
 export default function ResultDetails({ navigation }) {
-  const route = useRoute(); // Use useRoute hook to access the route prop
+  const route = useRoute();
+  const { Result } = route.params;
 
-  // Access the passed title from route.params
-  const { title } = route.params;
+  const mapViewRef = useRef(null);
+
   useEffect(() => {
     navigation.getParent().setOptions({ tabBarStyle: { display: "none" } });
     return () => {
       navigation.getParent().setOptions({ tabBarStyle: { display: "flex" } });
     };
-  }, []);
+  }, [navigation]);
+
+  const startLocation = {
+    latitude: parseFloat(Result.startPositionLatitude),
+    longitude: parseFloat(Result.startPositionLongitude),
+  };
+  const endLocation = {
+    latitude: parseFloat(Result.endPositionLatitude),
+    longitude: parseFloat(Result.endPositionLongitude),
+  };
+
+  useEffect(() => {
+    if (mapViewRef.current) {
+      mapViewRef.current.fitToCoordinates([startLocation, endLocation], {
+        edgePadding: { top: 50, right: 50, bottom: 50, left: 50 },
+        animated: true,
+      });
+    }
+  }, [startLocation, endLocation]);
+
   return (
-    <View style={{ flex: 1, backgroundColor: "white" }}>
-      <Image
-        source={maps}
-        style={{ width: wp(100), height: hp(55), resizeMode: "cover" }}
-      />
-      <StatusBar style={"light"} />
-      <SafeAreaView className="flex-row justify-between items-center w-full absolute">
+    <View style={{ flex: 1 }}>
+      <MapView
+        ref={mapViewRef}
+        style={{ flex: 1 }}
+        initialRegion={{
+          latitude: (startLocation.latitude + endLocation.latitude) / 2,
+          longitude: (startLocation.longitude + endLocation.longitude) / 2,
+          latitudeDelta:
+            Math.abs(startLocation.latitude - endLocation.latitude) * 1.5,
+          longitudeDelta:
+            Math.abs(startLocation.longitude - endLocation.longitude) * 1.5,
+        }}
+      >
+        <Marker coordinate={startLocation} title="Start" />
+        <Marker coordinate={endLocation} title="End" />
+        <MapViewDirections
+          origin={startLocation}
+          destination={endLocation}
+          apikey={"AIzaSyCSIIjbB9x_zvRGGGjnTlHS1qS8xXNGcxU"}
+          strokeWidth={3}
+          strokeColor="#FF0000"
+          onError={(errorMessage) =>
+            console.error("GMaps Directions Error: ", errorMessage)
+          }
+        />
+      </MapView>
+      <StatusBar style="light" />
+      <SafeAreaView
+        style={{
+          position: "absolute",
+          top: 0,
+          left: 0,
+          right: 0,
+          flexDirection: "row",
+          justifyContent: "space-between",
+          alignItems: "center",
+          paddingHorizontal: 16,
+        }}
+      >
         <TouchableOpacity
           onPress={() => navigation.goBack()}
-          className="p-1 rounded-lg m-4"
-          style={{ backgroundColor: "rgba(255,255,255, 0.5)" }}
+          style={{
+            padding: 10,
+            borderRadius: 5,
+            backgroundColor: "rgba(255, 255, 255, 0.5)",
+          }}
         >
-          <Feather name="arrow-left" size={wp(7)} color="white" />
+          <Feather name="arrow-left" size={24} color="white" />
         </TouchableOpacity>
       </SafeAreaView>
 
-      {/*Events description*/}
-
-      <View
-        style={{ borderTopLeftRadius: 20, borderTopRightRadius: 20 }}
-        className="px-5 flex flex-1 justify-between bg-white pt-8 -mt-14"
-      >
-        <ScrollView
-          showsVerticalScrollIndicator={false}
-          className="space-y-2 px-5"
-        >
-          <View className="mb-4 ">
-            <Text style={{ fontSize: wp(7) }} className="font-bold flex-1">
-              Marathon
-            </Text>
-            <Text
-              style={{ fontSize: wp(4), color: theme.text }}
-              className="flex-1"
-            >
-              Type
-            </Text>
-          </View>
-
-          <View className="flex justify-between mx-1 ">
-            <View className="flex-row space-x-2 items-start mb-4">
-              <Feather name="calendar" size={25} color={theme.text} />
-              <View className="flex space-y-2">
-                <Text
-                  style={{ fontSize: wp(4) }}
-                  className="font-bold text-neutral-700"
-                >
-                  Tues, 5 March 2024 at 12:30 PM
-                </Text>
-              </View>
-            </View>
-
-            <View className="flex-row space-x-2 items-start mb-4">
-              <Feather name="map-pin" size={24} color={theme.text} />
-              <View className="flex space-y-2">
-                <Text
-                  style={{ fontSize: wp(4) }}
-                  className="font-bold text-neutral-700"
-                >
-                  Golden Gate Work
-                </Text>
-              </View>
-            </View>
-          </View>
-
-          <View className="flex-row justify-between mx-1 mb-4">
-            {/*Distance*/}
-            <View className="flex-row space-x-2 items-start">
-              <View className="flex space-y-2">
-                <Text
-                  className="text-neutral-600 tracking-wide"
-                  style={{ fontSize: wp(4.5) }}
-                >
-                  Distance
-                </Text>
-                <Text className="font-bold text-neutral-700 text-center">
-                  10KM
-                </Text>
-              </View>
-            </View>
-            {/*Time*/}
-            <View className="flex-row space-x-2 items-start">
-              <View className="flex space-y-2">
-                <Text
-                  className="text-neutral-600 tracking-wide"
-                  style={{ fontSize: wp(4.5) }}
-                >
-                  Time
-                </Text>
-                <Text className="font-bold text-neutral-700 text-center">
-                  40:30
-                </Text>
-              </View>
-            </View>
-            {/*Time*/}
-            <View className="flex-row space-x-2 items-start">
-              <View className="flex space-y-2">
-                <Text
-                  className="text-neutral-600 tracking-wide"
-                  style={{ fontSize: wp(4.5) }}
-                >
-                  AVG Pace
-                </Text>
-                <Text className="font-bold text-neutral-700 text-center">
-                  4:30
-                </Text>
-              </View>
-            </View>
-          </View>
-
-          {/*Second row*/}
-
-          <View className="flex-row justify-between mx-1">
-            {/*Distance*/}
-            <View className="flex-row space-x-2 items-start">
-              <View className="flex space-y-2">
-                <Text
-                  className="text-neutral-600 tracking-wide"
-                  style={{ fontSize: wp(4.5) }}
-                >
-                  Distance
-                </Text>
-                <Text className="font-bold text-neutral-700 text-center">
-                  10KM
-                </Text>
-              </View>
-            </View>
-            {/*Time*/}
-            <View className="flex-row space-x-2 items-start">
-              <View className="flex space-y-2">
-                <Text
-                  className="text-neutral-600 tracking-wide"
-                  style={{ fontSize: wp(4.5) }}
-                >
-                  Time
-                </Text>
-                <Text className="font-bold text-neutral-700 text-center">
-                  40:30
-                </Text>
-              </View>
-            </View>
-            {/*Time*/}
-            <View className="flex-row space-x-2 items-start">
-              <View className="flex space-y-2">
-                <Text
-                  className="text-neutral-600 tracking-wide"
-                  style={{ fontSize: wp(4.5) }}
-                >
-                  AVG Pace
-                </Text>
-                <Text className="font-bold text-neutral-700 text-center">
-                  4:30
-                </Text>
-              </View>
-            </View>
-          </View>
-        </ScrollView>
-        <TouchableOpacity
-          style={{ backgroundColor: "#1C8FE3", height: wp(15), width: wp(50) }}
-          className="mb-6 mx-auto flex justify-center items-center rounded-xl"
-        >
-          <Text className="text-white font-bold" style={{ fontSize: wp(4.5) }}>
-            Download
+      <ScrollView style={{ flex: 1 }}>
+        <View style={{ padding: 20 }}>
+          <Text style={{ fontSize: 24, fontWeight: "bold", marginBottom: 10 }}>
+            {Result.eventName}
           </Text>
-        </TouchableOpacity>
-      </View>
+          <Text style={{ fontSize: 18, color: theme.text, marginBottom: 20 }}>
+            {Result.eventType}
+          </Text>
+          <View
+            style={{
+              flexDirection: "row",
+              alignItems: "center",
+              marginBottom: 20,
+            }}
+          >
+            <Feather name="calendar" size={24} color={theme.text} />
+            <Text style={{ marginLeft: 10 }}>
+              {Result.startDate.split("T")[0]} at{" "}
+              {Result.startTime.substring(0, 5)}
+            </Text>
+          </View>
+          <View
+            style={{
+              flexDirection: "row",
+              alignItems: "center",
+              marginBottom: 20,
+            }}
+          >
+            <Feather name="map-pin" size={24} color={theme.text} />
+            <Text style={{ marginLeft: 10 }}>Sulzbach an der Murr</Text>
+          </View>
+          <View
+            style={{
+              flexDirection: "row",
+              justifyContent: "space-between",
+              marginBottom: 20,
+            }}
+          >
+            <View>
+              <Text style={{ fontSize: 18 }}>Calories</Text>
+              <Text style={{ fontSize: 24, fontWeight: "bold" }}>
+                {Result.calories.toFixed(2)}
+              </Text>
+            </View>
+            <View>
+              <Text style={{ fontSize: 18 }}>Time</Text>
+              <Text style={{ fontSize: 24, fontWeight: "bold" }}>
+                {Result.time.substring(0, 5)}
+              </Text>
+            </View>
+            <View>
+              <Text style={{ fontSize: 18 }}>AVG Pace</Text>
+              <Text style={{ fontSize: 24, fontWeight: "bold" }}>
+                {Result.pace.substring(3, 8)}
+              </Text>
+            </View>
+          </View>
+        </View>
+      </ScrollView>
     </View>
   );
 }
