@@ -1,47 +1,82 @@
-import { FlatList, View, ScrollView, Animated } from "react-native";
-import React from "react";
+import React, { useState } from "react";
+import { FlatList, ActivityIndicator, View, Platform } from "react-native";
+import { useFocusEffect, useNavigation } from "@react-navigation/native";
 import ResultsHeader from "./ResultsHeader";
 import ResultsElement from "./ResultsElement";
-import { useNavigation } from "@react-navigation/native";
 import { SafeAreaView } from "react-native-safe-area-context";
+import AsyncStorage from "@react-native-async-storage/async-storage";
+import api from "../../Api/api";
+import { theme } from "../../Theme";
 
 export default function ResultsScreen() {
-  {
-    /*After that change it with the data from DB*/
-  }
+  const [data, setData] = useState([]);
+  const [loading, setLoading] = useState(true);
   const navigation = useNavigation();
 
-  const handlePress = (title) => {
-    navigation.navigate("Details", { title });
-  };
-  const data = [
-    { id: "1", title: "Item 1" },
-    { id: "2", title: "Item 2" },
-    { id: "3", title: "Item 3" },
-    { id: "4", title: "Item 4" },
-    { id: "5", title: "Item 5" },
-    { id: "6", title: "Item 6" },
-    { id: "7", title: "Item 7" },
-    { id: "8", title: "Item 8" },
-    { id: "9", title: "Item 9" },
-    { id: "10", title: "Item 10" },
-  ];
-  const renderItem = ({ item }) => (
-    <ResultsElement title={item.title} handlePress={handlePress} />
+  useFocusEffect(
+    React.useCallback(() => {
+      const fetchEvents = async () => {
+        try {
+          const token = await AsyncStorage.getItem("Token");
+          const email = await AsyncStorage.getItem("Email");
+          if (token) {
+            const response = await api.get(
+              `ParticipantRuns/GetRunsResult?email=${email}`,
+              {
+                headers: { Authorization: `Bearer ${token}` },
+              }
+            );
+            console.log(response.data);
+            setData(response.data);
+          } else {
+            console.error("Token not found in AsyncStorage");
+          }
+        } catch (error) {
+          console.error("Error fetching events:", error);
+        } finally {
+          setLoading(false);
+        }
+      };
+
+      fetchEvents();
+    }, [])
   );
+
+  const handlePress = (item) => {
+    navigation.navigate("Details", { Result: item });
+  };
+
+  const renderItem = ({ item }) => (
+    <ResultsElement data={item} handlePress={() => handlePress(item)} />
+  );
+
   const topMargin =
     Platform.OS === "ios" ? { marginTop: 3 } : { marginTop: 15 };
 
+  if (loading) {
+    return (
+      <View style={[styles.loadingContainer, topMargin]}>
+        <ActivityIndicator size="large" color={theme.text} />
+      </View>
+    );
+  }
+
   return (
-    <SafeAreaView className="flex-1 bg-white">
-      <ResultsHeader />
-      {/*This need to be changed when I fetch the data from the database*/}
+    <View style={{ flex: 1, backgroundColor: "white" }}>
       <FlatList
         showsVerticalScrollIndicator={false}
         data={data}
         renderItem={renderItem}
         keyExtractor={(item) => item.id}
       />
-    </SafeAreaView>
+    </View>
   );
 }
+
+const styles = {
+  loadingContainer: {
+    flex: 1,
+    justifyContent: "center",
+    alignItems: "center",
+  },
+};
